LINUX: Introduce '-atsys vfs'
Currently, the way we do an @sys lookup looks something like this:

    afs_linux_lookup(struct dentry *dp, ...)
    {
        inode = afs_lookup(dp->d_name, ...);
        return d_splice_alias(inode, dp);
    }

    afs_lookup(char *comp, ...)
    {
        foreach (sysname) {
            name = substitute_sysname(comp, sysname);
            inode = afs_try_lookup(name);
            if (inode) {
                return inode;
            }
        }
    }

That is, we call afs_lookup() with the name given to us by the Linux
VFS, and afs_lookup() internally tries substituting @sys with the
various sysnames that we have during the lookup. afs_linux_lookup()
then tries to attach the given dentry to the inode returned by
afs_lookup(), and returns the result.

This generally works fine, but this means that we can easily have 2
dentries for the same directory (if the expanded @sys is a directory):
one for '@sys' (if someone tries to access '@sys'), and one for e.g.
'amd64_linux26' (if someone tries to access 'amd64_linux26' directly,
without going through '@sys'). We have some workarounds in our Linux
integration to deal with this (for e.g. multiple mountpoints for the
same volume), but it would be better if we didn't need to have
multiple dentries for the same inode at all.

So to avoid this, we can take a different approach. Instead of
afs_lookup() performing the @sys substitution internally, we try to
substitute @sys with a sysname string ourselves, and then go through
the Linux VFS for the lookup on the substituted string. The resulting
code looks something like this:

    afs_linux_lookup(struct dentry *dp, ...)
    {
	if (ends_in_atsys) {
	    foreach (sysname) {
		name = substitute_sysname(dp->d_name, sysname);
		dentry = afs_lookup_noperm(name, ...); /* call into Linux VFS */
		if (dentry) {
		    return dentry;
		}
	    }
	    return ENOENT;
	}
        inode = afs_lookup(dp->d_name, ...);
        return d_splice_alias(inode, dp);
    }

Note that for the code path where we do the @sys substitution
ourselves (and do not call afs_lookup()), we don't add the given
dentry to the looked-up inode at all; we just return a different
existing dentry. Since this is guaranteed to be a dentry for the same
directory as we were already trying to do the lookup in, this should
not cause issues with the Linux VFS like those mentioned in commit
dda3ea5f ("Linux: Make dir dentry aliases act like symlinks").

This means that we will not cache a dentry for the literal string
'@sys', so Linux will call afs_linux_lookup() on every lookup request
for '@sys'. But this should not be bad for performance, since for the
common @sys case, Linux will service the afs_lookup_noperm() request
(aka lookup_one_len()) by looking in the dcache, so we perform very
similar to if the '@sys' entry was cached in the dcache anyway.

To avoid killing performance for this code path, we need to avoid
taking GLOCK while doing the @sys substitution. In order to do so,
we examine the sysname list under afs_sysname_lock, without taking
GLOCK.

Also note that when we do @sys expansion in this new way, we must make
sure to not do @sys expansion in afs_lookup(). Otherwise, if any of
the sub-lookups in afs_linux_lookup() failed for a transient reason,
we could do an old-style @sys expansion inside the later afs_lookup()
call, and so our @sys paths may be mixed between the old and new way.

While this new behavior should be better in almost every way, it does
result in user-visible changes to @sys paths. Because we don't make a
dentry for the string '@sys', things like getcwd() now always return the
expanded path name, where before we sometimes returned the '@sys' path,
and sometimes the expanded path. We also never try to lookup a path name
with the unusual literal name '@sys'. Because this is a user-visible
change, implement this new behavior in a new -atsys mode, called 'vfs',
and (at least for now), the default behavior is unchanged.

Restrict the new functionality to Linux kernel 3.10+, just to avoid
risking any breakage or edge cases with older kernels.

This commit introduces the constant AFS_ATSYS_VFS (to represent the new
-atsys mode), and the preprocessor symbol AFS_ATSYS_VFS_ENV (to indicate
we have actually implemented the '-atsys vfs' mode). AFS_ATSYS_VFS is
only defined on LINUX for now, since no other platform implements the
option. Most cross-platform code checks if AFS_ATSYS_VFS is defined to
see if we can process the option at all, but we check for
AFS_ATSYS_VFS_ENV to see if the option can actually be turned on.

This commit also moves the AFS_EQ_ATSYS macro from
VNOPS/afs_vnop_lookup.c to afs.h, just so both code paths can use the
same macro.

Change-Id: I86493cb14f61d7354afd835e4d8ce99674210252
Reviewed-on: https://gerrit.openafs.org/13300
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>